+++ package/xag/nart_ver_4_9_844/src/devlib/ar9287/Ar9287NoiseFloor.c
@@ -4,7 +4,7 @@ int Ar9287NoiseFloorFetch(int *nfc, int *nfe, int nfn)
 {
 	unsigned long mask;
 	int it;
-	unsigned long address;
+	unsigned int address;
 	int high, low;
 	//
 	// noise floor values are signed. this computes a mask to extend the sign bit
@@ -23,26 +23,26 @@ int Ar9287NoiseFloorFetch(int *nfc, int *nfe, int nfn)
 	// fetch the answers
 	//
 #if 1
-    FieldRead("CCA.bb_minCCApwr", (unsigned long *)&nfc[0]);
+    FieldRead("CCA.bb_minCCApwr", (unsigned int *)&nfc[0]);
 	nfc[0]|=mask;
 	nfc[1]=0;
 	nfc[2]=0;
-	FieldRead("AGC_EXT.bb_minCCApwr_ext", (unsigned long *)&nfe[0]);
+	FieldRead("AGC_EXT.bb_minCCApwr_ext", (unsigned int *)&nfe[0]);
 	nfe[0]|=mask;
 	nfe[1]=0;
 	nfe[2]=0;
 #else
-	FieldRead("BB_cca_b0.minCCApwr_0", (unsigned long *)&nfc[0]);
+	FieldRead("BB_cca_b0.minCCApwr_0", (unsigned int *)&nfc[0]);
 	nfc[0]|=mask;
-	FieldRead("BB_cca_b1.minCCApwr_1", (unsigned long *)&nfc[1]);   
+	FieldRead("BB_cca_b1.minCCApwr_1", (unsigned int *)&nfc[1]);   
 	nfc[1]|=mask;
-	FieldRead("BB_cca_b2.minCCApwr_2", (unsigned long *)&nfc[2]);
+	FieldRead("BB_cca_b2.minCCApwr_2", (unsigned int *)&nfc[2]);
 	nfc[2]|=mask;
-	FieldRead("BB_ext_chan_pwr_thr_2_b0.minCCApwr_ext_0", (unsigned long *)&nfe[0]);
+	FieldRead("BB_ext_chan_pwr_thr_2_b0.minCCApwr_ext_0", (unsigned int *)&nfe[0]);
 	nfe[0]|=mask;
-	FieldRead("BB_ext_chan_pwr_thr_2_b1.minCCApwr_ext_1", (unsigned long *)&nfe[1]);
+	FieldRead("BB_ext_chan_pwr_thr_2_b1.minCCApwr_ext_1", (unsigned int *)&nfe[1]);
 	nfe[1]|=mask;
-	FieldRead("BB_ext_chan_pwr_thr_2_b2.minCCApwr_ext_2", (unsigned long *)&nfe[2]);
+	FieldRead("BB_ext_chan_pwr_thr_2_b2.minCCApwr_ext_2", (unsigned int *)&nfe[2]);
 	nfe[2]|=mask;
 #endif
 	return 0;    
@@ -71,7 +71,7 @@ int Ar9287NoiseFloorLoad(int *nfc, int *nfe, int nfn)
 
 int Ar9287NoiseFloorReady(void)
 {
-	int ready;
+	unsigned int ready;
 
 	FieldRead("AGC_CTL2.bb_do_noisefloor",&ready);
 	return ready;
@@ -100,7 +100,7 @@ void Ar9287_REG_do_noisefloor_Write(int value)
     FieldWrite("AGC_CTL2.bb_do_noisefloor",value);
 }
 
-void Ar9287_REG_do_noisefloor_Read(unsigned long *value)
+void Ar9287_REG_do_noisefloor_Read(unsigned int *value)
 {
     FieldRead("AGC_CTL2.bb_do_noisefloor",value);
 }
